Make use of builder pattern optional
Now the builder patterns are only generated if the "builder" feature is enabled
(which is the default). Disabling it avoids some heavy dependencies for implementing the procedural macros, which can save a lot of compile time.

This is technically a breaking change for anyone who builds zxcvbn with
`no-default-features = true` and uses the builder patterns. (This should not be
the case often, because the zxcvbn did not have default features before.)